Eliminate ufc-crypt nonce types.
ufc_long, long32, and long64 are all replaced with the underlying
stdint.h types, and _UFC_32_ and _UFC_64_ are replaced with a single
macro UFC_USE_64BIT which is always defined but may be true or false.
ufc-crypt.h is deleted and xcrypt-private.h no longer includes
crypt-private.h.  Some other minor cruft is also eliminated.

	* src/bigcrypt.c: Don't include crypt.h, ufc-crypt.h, or
	crypt-private.h.  Delete ancient commented-out debugging printfs.
	* src/crypt-private.h: Don't include crypt.h or ufc-crypt.h; do
	include stdint.h.  Forward-declare struct crypt_data here.
	Define UFC_USE_64BIT here.  Replace ufc_long with uint_fast32_t.
	* src/xcrypt-private.h: Don't include crypt-private.h.

	* src/crypt-entry.c, src/crypt.c, src/crypt_util.c:
	Don't include ufc-crypt.h.  Include xcrypt-private.h and
	crypt-private.h above system headers.
	Replace ufc_long with uint_fast32_t, long32 with uint32_t, and
	long64 with uint64_t.  Change all #ifdefs keyed on _UFC_32_ and/or
	_UFC_64_ to use UFC_USE_64BIT instead.