Eliminate all internal requirements for aligned buffers.
The MD5, SHA256, and SHA512 implementations all internally tried to
read their input buffers a uint32_t or uint64_t at a time, with
associated hacks to cater to CPUs that don't do efficient unaligned
word reads, and to ensure the algorithmically expected endianness of
word quantities.  This kind of performance hack is appropriate for
hashing of bulk data, but unnecessary for password hashes, which are
deliberately made slow anyway.  Strip all of that out and use
idiomatic sequences of byte reads instead, encapsulated in the new
header byteorder.h; spot-checking indicates that GCC 6 does convert
these to optimal instruction sequences for x86, at least.

This also eliminates almost all of the places where 'alloca' is used
with a caller-controlled buffer size.  (A few remain.)

In passing, we clean up #includes in the affected files a bunch.
The only remaining use of a non-ISO C standard header, in the entire
library, is <sys/types.h> included from crypt.h.  This is needed for
the definitions of __THROW and __nonnull, and will be addressed in a
future patch.

	* src/byteorder.h: New file.
	* src/test-byteorder.c: New test program.

	* src/md5.h, src/sha256.h, src/sha512.h:
	Minimize includes.  Remove unused definitions.  Do not increase
	alignment of scratch buffers.  Use uint64_t instead of 2x uint32_t
	when possible.  Remove __THROW annotations, as these headers will
	never be used from C++.  Improve commentary.

	* src/md5.c, src/sha256.c, src/sha512.c:
	Minimize includes.  Use byteorder.h functions to read and write
	data, instead of SWAP and pointers to uint32_t/uint64_t.  Use
	64-bit arithmetic instead of 2x32-bit arithmetic.  No need to
	break large blocks of input into chunks for alignment's sake.

	* src/md5-crypt.c, src/sha256-crypt.c, src/sha512-crypt.c:
	Minimize includes.  Do not increase alignment of scratch buffers.
	Do not copy input data to align it.

	* src/test-sha512.c: Minimize includes. Print details on failure.
	* src/test-sha256.c: Likewise.  Remove a duplicate test vector.
	* src/test-md5.c: Likewise.  Add two more test vectors from NIST.

	* src/Makefile.am: Add byteorder.h and test-byteorder.
	* .gitignore: Update for new test.