Eliminate remaining uses of alloca and mempcpy.
The last four uses of alloca were in sha256-crypt.c and
sha512-crypt.c, to allocate space for intermediate byte sequences
consisting of intermediate digests repeated to the length of the key
or salt.  Instead, store just the intermediate digests, and "recycle"
them (as the R people call it) by calling sha{256,512}_process_bytes
in a loop.  This also eliminates all uses of the nonstandard function
'mempcpy'.

Also delete some no-longer-needed local variables; this could've been
in the previous patch, but whatever.

	* src/sha256-crypt.c (sha256_process_recycled_bytes): New function.
	(_xcrypt_crypt_sha256_rn): Store only the fixed-length digest for
	'p_bytes' and 's_bytes'; use sha256_process_recycled_bytes to get
	the same effect as repeating them in memory to the length of the
	password or salt.  Remove variables 'copied_key' and
	'copied_salt', unnecessary since previous patch.
	* src/sha512-crypt.c: Similarly.

	* configure.ac: Don't call AC_FUNC_ALLOCA.
	* m4/zw_automodern.m4: Don't put stddef.h in ac_includes_default.