Don't use bare __attribute__.
The only current use of __attribute__ is to suppress unused-argument
warnings in a couple places.  Add an ARG_UNUSED macro to libc-symbols.h
which expands to either the GNU __attribute__, or the C++ convention
of omitting the parameter name, or to nothing special, as appropriate.

Also remove the currently-unused 'weak_alias' from that file.
Something like that is going to come back later when we tackle
symbol-version compatibility, but for now it's cruft.

	* src/libc-symbols.h (weak_alias, _weak_alias): Delete.
	(ARG_UNUSED): New macro.
	* src/des-crypt.c, src/xcrypt-gensalt.c: Use ARG_UNUSED
	instead of bare __attribute__((unused)).