Reorganize source tree (4/n): more name prefixes.
Files that just implement a cryptographic primitive are now named
alg-*, and files that implement the high-level interface and entire
hash modules are now named crypt-*.  The xcrypt- prefix is gone.
Tests are renamed to match.  libc-symbols.h, which no longer has
anything to do with libc, is renamed crypt-symbols.h.

Also, the library built by this source tree is named libcrypt again.
The ultimate goal, after all, is to make this a drop-in,
binary-upward-compatible replacement for glibc's libcrypt.  (It's
still libcrypt.so.2, though.  We're not there yet.)

Take the opportunity to rationalize the organization of stuff in
src/Makefile.am.