Reintroduce semi-automated namespace cleanliness.
All of the internal symbols that can't be 'static' in a single file
are now renamed with a _crypt_ prefix by macros in crypt-symbols.h.
The list of these macros must be maintained by hand, but there is a
test program (test-symbols.sh) that verifies it.  Essentially, all
global symbols in the static library (libcrypt.a) must either appear
in a "global:" block in libcrypt.map, or they must begin with _crypt_.
It also checks that every "global:" symbol declared in libcrypt.map is
exported from the static library.

This reminded me that two API symbols should never have existed in the
first place (bigcrypt_r, crypt_gensalt_r) and one had gotten dropped
by accident (fcrypt).

Due to limitations in Automake, invoking the new test script by hand
is a little awkward (it takes arguments from the environment) and the
C programs used only for testing are now part of check_PROGRAMS, not
noinst_PROGRAMS, which means they are no longer built by 'make all'.

	* crypt-obsolete.h (bigcrypt_r): Remove declaration.
	* crypt.h (crypt_gensalt_r): Remove declaration.
	* libcrypt.map (crypt_gensalt_r, bigcrypt_r): Remove.
	* crypt-symbols.h: Correct multiple-inclusion guard.
	Add macros that rename all internal global symbols.
	Add new symbol-hacking macro 'strong_alias'.
	* crypt.c (crypt_gensalt_r): Delete.
	(bigcrypt_r): Fold body into bigcrypt and delete.
	(fcrypt): Define as strong alias of crypt.
	* test-gensalt.c: Use crypt_gensalt_rn, not crypt_gensalt_r.

	* test-symbols.sh: New test.
	* Makefile.am (noinst_PROGRAMS): Just alg-des-mktables.
	(check_PROGRAMS): All C tests are here now.
	(TESTS): Add test-symbols.sh.
	(AM_TESTS_ENVIRONMENT): Provide arguments for test-symbols.sh.