Add configure options to disable obsolete APIs and/or weak hashes.
If the library is configured with --disable-obsolete-api, the functions
encrypt, encrypt_r, setkey, setkey_r, bigcrypt, and fcrypt will not
be included in the library, not even as dynamic-only symbols.  This
also changes the SONAME of the library from libcrypt.so.1 (compatible
with glibc) to libcrypt.so.2, and removes all of the compatibility
symbol versions (GLIBC_*) from the supported APIs.

If the library is configured with --disable-weak-hashes, the code for
DES and MD5 password hashes will not be included in the library, which
means the crypt() functions will not be able to generate or verify
password hashes in these formats.  This option implies --disable-obsolete-api.

	* configure.ac: Add AC_PROG_CPP.  Move zw_SIMPLE_ENABLE_WARNINGS
	above LT_INIT.
	(--disable-obsolete-api, --disable-weak-hashes): New configure
	options.
	(ENABLE_OBSOLETE_API, ENABLE_WEAK_HASHES): New conditional flags,
	available in Automake and as preprocessor macros.
	* Makefile.am: Conditionally include support for the obsolete APIs
	and the weak hashes.
	(libcrypt.map): Is now a generated file.
	(EXTRA_libcrypt_la_DEPENDENCIES): Mention libcrypt.map.
	(libcrypt_la_VERSION): Make the library SONAME conditional on
	whether obsolete APIs are enabled.
	* crypt.c: Compile the table entries for DES and MD5 hash prefixes
	only when ENABLE_WEAK_HASHES is true.  Compile fcrypt and bigcrypt
	only when ENABLE_OBSOLETE_API is true.
	* test-gensalt.c: Compile the table entries for DES and MD5 hash
	prefixes only when ENABLE_WEAK_HASHES is true.  Do not crash
	if crypt_gensalt_rn returns NULL.
	* .gitignore: Update.