Data buffers for MD5 and SHA
As was previously done for DES, move all of the sensitive intermediate
data used by the MD5, SHA256, and SHA512 hashes into a single buffer
allocated by the caller, along with the output.  This makes it easier
to be sure that all of it gets erased reliably at the end of the
calculation, and will also facilitate getting it off the stack.
Bcrypt still needs the same treatment, but that's going to be more
involved.

In order to do this cleanly we need to use C11's _Alignof or a
fallback, so have configure detect that.  It also detects _Alignas,
which we don't need right this instant but I suspect we will in the
future.  Unfortunately, this means we are actually using a C11
feature, and therefore cannot use -Wc99-c11-compat.

	* m4/zw_alignment.m4: New Autoconf macros to detect _Alignof and
	_Alignas, or equivalent GNU extensions.
	* configure.ac: Use them.
	* m4/zw_simple_warnings.m4: Remove -Wc99-c11-compat.

	* alg-md5.h (struct md5_ctx): Add correct_words.
	* alg-md5.h (md5_process_bytes): correct_words moved into md5_ctx.
	* crypt-des.c: Use alignof instead of guessing the necessary
	alignment for struct des_ctx.
	* crypt-md5.c (md5_hash_length): Make into a preprocesor constant,
	MD5_HASH_LENGTH.
	(struct md5_buffer, md5_get_ctx, md5_get_alt_ctx)
	(md5_wipe_intermediate_data): New.
	(crypt_md5_rn): All scratch aggregates allocated from an
	md5_buffer overlaid on the 'data' argument, instead of
	being stack locals.
	* crypt-sha256.c, crypt-sha512.c: Similarly.