Data buffers for bcrypt
Move as much as possible of the sensitive intermediate data used by
bcrypt into a buffer allocated by the caller.  This is finickier
than for the other hashes because the bcrypt code is more complicated;
in particular, because BF_crypt no longer takes a simple pointer to
where its output will be written, additional shuffling-around of data
inside crypt_bcrypt_rn is necessary to accommodate the self-test.

As a consequence of this change, it no longer makes sense for
test-crypt-bcrypt to pass "output" buffers of varying sizes to
crypt_rn.