Implement precise symbol-version compatibility with GNU libc.
Up to this point, all of the symbols that existed in GNU libc's
libcrypt have been defined with a compatibility version of GLIBC_2.0.
This is inaccurate, because GNU libc has a minimum symbol version for
each supported CPU and OS.  For instance, the aarch64 port's minimum
version is GLIBC_2.17, so all symbols older than the 2.17 release will
be tagged GLIBC_2.17 in the compiled libraries, even if their Versions
file gives them an older version.

An additional complication is that the OpenWall crypt_* extensions
were never officially adopted by glibc upstream, but at least two
Linux distributions (OpenWall and ALT) patched them in -- with
different symbol versions.

This patch implements glibc-compatible minimum symbol versions for all
of the Linux ports currently present in glibc upstream.  We assume
other operating systems do not have a glibc libcrypt to be compatible
with, so we set --disable-obsolete-api for them, rendering the entire
mess moot.  (The known exception is of course the Hurd, but it's my
impression that binary backward compatibility is not a priority for
the Hurd yet, so --disable-obsolete-api is probably the Right Thing
for them too.)

	* libcrypt.map.in: Restructure into a tabular format that's
	easy to parse in awk.  Record all known compatibility symbol
	versions, not just GLIBC_2.0.
	* libcrypt.minver: New file recording the minimum symbol version
	used by glibc on all (current) Linux ports.
	* gen-map.awk, gen-vers.awk: New build scripts.

	* test-symbols-compat: New test, looks for an installed
	libcrypt.so.1 and compares its symbol versions with ours.
	* test-symbols.sh: Rename test-symbols-static; update for new
	format of libcrypt.map.in and re-activate the "missing symbols"
	test.

	* crypt-symbols.h: Include crypt-symbol-vers.h when IN_LIBCRYPT
	is defined.  Completely rewrite the symbol versioning macros so
	all of the complexity, not just most of it, is hidden in the
	generated code.

	* crypt.c, crypt-static.c, crypt-des-obsolete.c: Update for
	new symbol versioning macros.

	* configure.ac: Add AC_CANONICAL_HOST and AC_PROG_AWK.
	Improve help texts for --disable-obsolete-api and
	--disable-weak-hashes.
	Parse libcrypt.minver and set SYMVER_MIN and SYMVER_FLOOR
	substitution variables.
	Add AM_CONDITIONAL for ENABLE_STATIC.
	* Makefile.am: Define preprocessor macro IN_LIBCRYPT for all
	objects built into libcrypt.a/libcrypt.so.  Give all such objects
	a dependency on crypt-symbol-vers.h (this file is generated, so
	autodependency generation can't handle it).
	(libcrypt.map): Generate from libcrypt.map.in
	using gen-map.awk.
	(crypt-symbol-vers.h): Generate from libcrypt.map.in
	using gen-vers.awk.
	(TESTS): Only run test-symbols-static if ENABLE_STATIC.
	Run test-symbols-compat if ENABLE_OBSOLETE_API.
	(test_des_obsolete_LDADD): Link with libcrypt.la, not individual
	object files.
	* .gitignore: Update.