Shrink crypt_data and increase CRYPT_(GENSALT_)OUTPUT_SIZE.
struct crypt_data reserved enough space for UFC-DES - a little more
than 128kB.  We are no longer using UFC-DES, and none of the
currently-supported algorithms require nearly that much space.  (Some
algorithms that we may want to support in the future, such as scrypt
and Argon2, require *enormously* more memory - gigabytes - but that
will require a new interface, so there is no point worrying about it
now.  We reserve generous space for future expansion;
CRYPT_OUTPUT_SIZE and CRYPT_GENSALT_OUTPUT_SIZE are increased to 384
and 192 bytes respectively, and the size of crypt_data only drops to
32kB, of which the first 2kB is a documented interface, the next 8kB
is supplied to algorithm modules as scratch space (no algorithm
currently requires more than about half of this), and the remaining
22kB will be used by the patch coming up next.

This is the first of two planned API breaks.  After this change, it is
definitely true that programs compiled against this libcrypt will not
work correctly with glibc's libcrypt, because their crypt_data will be
too small.  However, programs compiled with glibc's libcrypt will
still work fine, because their crypt_data is larger than required and
this library doesn't actually pay any attention to its contents.  (The
"initialized" field is still documented, because that requirement
might come back in the future, but we don't look at it now.)  The
symbol versioning is already consistent with this -- existing binaries
use GLIBC_x.y symbols which we provide, new ones use XCRYPT_x.y
symbols which glibc doesn't provide.

There are a number of internal cleanups in this patch as well, since
we had to touch every algorithm module anyway: the algorithm modules
are no longer responsible for aligning their scratch space or for
erasing it after they are done with it, internal terminology has been
made consistent throughout ("phrase" or "passphrase" instead of
"password" or "key"; "setting" for the full setting string, "salt"
only for the actual salt component of the setting string; "rbytes" and
"nrbytes" instead of "input" and "size" for the random-byte string fed
to crypt_gensalt), and we take a few more steps in the direction of
consistently using unsigned char for internal processing.

	* crypt-base.h (CRYPT_OUTPUT_SIZE): Increase to 384.
	(CRYPT_GENSALT_OUTPUT_SIZE): Increase to 192.
	(CRYPT_MAX_PASSPHRASE_SIZE): New macro (value 512).
	(struct crypt_data): Contents completely changed.  Now only 32kB.
	(crypt, crypt_r, crypt_ra, crypt_rn)
	(crypt_gensalt, crypt_gensalt_rn, crypt_gensalt_rn):
	Improve documentation and formal parameter names.

	* crypt.c: Improve formal parameter names throughout.
	(struct crypt_internal, crypt_fn, gensalt_fn): New types.
	(get_internal, call_crypt_fn): New functions.
	(make_failure_token): If there is not enough space for the
	full failure token, write as much of it as possible.
	(do_crypt_rn, do_crypt_ra): Use call_crypt_fn.
	(crypt_gensalt_rn): Make input/output buffer size checks more
	thorough.  Use make_failure_token.

	* crypt-private.h: Don't include stddef.h.
	(ALG_SPECIFIC_SIZE): New macro.
	(all function prototypes): Update.

	* crypt-bcrypt.c: Use unsigned char more.  Add static_asserts
	that CRYPT_OUTPUT_SIZE and ALG_SPECIFIC_SIZE are big enough.
	Remove the 'output' field from struct BF_buffer.
	(BF_get_data, BF_wipe_intermediate_data): Delete.
	(crypt_bcrypt_rn, BF_gensalt): Update to match internal API change.
	* crypt-des.c, crypt-md5.c, crypt-sha256.c, crypt-sha512.c:
	Similarly.

	* m4/zw_alignment.m4 (zw_C_MAX_ALIGN_T): New macro.
	* configure.ac: Call it.
	* crypt-symbols.h: Add fallback definition of max_align_t.
	Include stddef.h and stdint.h here.  Provide definitions of
	MIN and MAX here.