Use swapcontext for thorough erasure of sensitive intermediate data.
We are already segregating most of the sensitive data used by the
crypt algorithm modules in a scratch area that gets erased after
hashing is complete, but there may be sensitive data left on the stack
or in registers.  We can arrange to erase this data as well by using
the dusty old 'getcontext', 'makecontext', and 'swapcontext' functions
to run the hashing algorithm using a special stack and register save
area whose locations we know.  This is what we needed the extra 22kB
in struct crypt_data for.  Because these functions are dusty and old,
we have to do a bunch of nasty pointer-swizzling on LP64 platforms,
but the extra security should be worth the hassle.

Because we are now doing nontrivial portability checks in the
configure script, also change zw_automodern.m4 to redefine
AC_INCLUDES_DEFAULT with its actual definition in 2.70-to-be, rather
than providing no default header checks at all.  (In particular, this
means the test for getcontext will work correctly on systems where
ucontext.h requires you to include sys/types.h first.  The *code*
does not handle this case yet, but will shortly.)

	* configure.ac: Check for ucontext.h and the getcontext family of
	functions.
	* crypt.c: If USE_SWAPCONTEXT is defined, reserve space for an
	inner stack and a ucontext_t in struct crypt_internal, and call
	the hash algorithm on that stack so that anything it puts on the
	stack can be erased afterward.

	* m4/zw_automodern.m4 (_AC_INCLUDES_DEFAULT_REQUIREMENTS):
	Provide the same set of inclusions and header checks that
	the actual 2.70 AC_INCLUDES_DEFAULT will.