Grab bag of header-related cleanups.
Stop including crypt-symbols.h implicitly. It's a little too much
_spukhafte Fernwirkung_ for my taste, and it makes Emacs have trouble
parsing the compiler's error messages.

Also, rename crypt-symbols.h to crypt-port.h as this is more
consistent with its current function, and have it include stdbool.h,
stddef.h, and stdint.h for everyone; have configure bomb out if the
C99 headers we use are not available; and normalize the order of
inclusion in all files.

The value of CPPFLAGS passed to test scripts now includes every last
one of Automake's CPPFLAGS-fragment variables.