Restore compilability with clang 3.8.
The reshuffle of struct crypt_data caused clang to be unhappy about
alignment again, which was really just a symptom of crypt-des-obsolete.c
not having been properly updated for the new layout.  setkey_r and
encrypt_r should stash the des_ctx in the "internal" area, like
crypt_r does, but setkey and encrypt don't bother allocating the full
crypt_data, so they need to bypass that.

Also eliminate an incompletely initialized compound literal in
a static_assert.

	* crypt-base.h (CRYPT_DATA_RESERVED_SIZE)
	(CRYPT_DATA_INTERNAL_SIZE): New macros.
	* crypt.c: Use CRYPT_DATA_INTERNAL_SIZE.
	* crypt-des-obsolete.c (get_des_ctx, do_setkey_r, do_encrypt_r):
	New functions.
	(setkey_r, encrypt_r): Refactor to use get_des_ctx.
	(setkey, encrypt): Refactor to avoid get_des_ctx.