Be compatible with SuSE's version of the bcrypt-in-glibc patches.
SuSE includes the OpenWall patchset in its libcrypt with GLIBC_2.0
symbol versions, but it also ships a library called 'libowcrypt' that
contains only the crypt_gensalt_* functions, under another symbol
version (OW_CRYPT_1.0).  In ENABLE_OBSOLETE_API mode, therefore, add
another set of version aliases to the crypt_gensalt_* functions, and
symlink libowcrypt.{a,so,so.1} to libcrypt.{a,so,so.1} as appropriate.

SuSE also includes "gnu-crypt.h" and "ow-crypt.h" variations on
crypt.h.  We will not be providing these for now, as any program that
uses them almost certainly has #ifdeffage to fall back to regular
crypt.h.

	* Makefile.am (install-exec-hook, uninstall-hook): New rules
	to install and uninstall compatibility symlinks for SuSE's
	libowcrypt.
	* configure.ac: Add AC_PROG_LN_S and an AM_CONDITIONAL for
	ENABLE_SHARED, used by the above.
	* libcrypt.map.in (crypt_gensalt, crypt_gensalt_rn)
	(crypt_gensalt_ra): Add additional OW_CRYPT_1.0 compatibility
	version.