Add default algorithm choice and randomization to crypt_gensalt.
If the 'prefix' argument to a crypt_gensalt function is NULL, it will
now choose the best currently-supported algorithm.  If the 'rbytes'
argument to a crypt_gensalt function is NULL, it will now get random
bytes from the operating system.

The upshot of all that is you can do

    char *setting = crypt_gensalt_ra (0, 0, 0, 0);

and get a sensible set of defaults.  CAUTION: supplying an *empty string*
as the prefix argument ("", rather than 0) will cause it to generate a
setting string that selects DES, which is not what you want.
(Unfortunately, we are stuck with this behavior until we can make
--disable-weak-hashes the default, which may never happen.)

This is the second of two planned backward compatibility breaks for
libxcrypt; programs that do the above will not work correctly with
glibc's libcrypt (even if it supplies crypt_gensalt).

Note: randombytes.c is full of portability ifdeffage, but the only tested
code paths in there are the syscall(SYS_getrandom) case and the final
fallback to /dev/urandom.  Tweaks will probably be necessary upon wider
testing.

	* crypt.c (crypt_gensalt_rn): If prefix is 0, use the prefix for
	the first entry in the tagged_hashes array, which was already
	documented as the proper choice for new hashes.  If rbytes is 0,
	call get_random_bytes and use its output instead.
	* test-gensalt.c: Test behavior of passing 0 as prefix and/or
	rbytes arguments to crypt_gensalt.

	* crypt-base.h (crypt_gensalt, crypt_gensalt_rn)
	(crypt_gensalt_ra): Document new semantics of passing a null
	pointer for prefix or rbytes argument; adjust __nonnull
	annotations to allow these arguments to be null.

	* randombytes.c: New file.
	* crypt-port.h, crypt-private.h: New internal function get_random_bytes.
	* Makefile.am (libcrypt_la_SOURCES): Add randombytes.c.
	* configure.ac: Add checks for fcntl.h, sys/random.h, sys/syscall.h,
	arc4random_buf, getentropy, getrandom, and syscall.