Stop using VLA function parameters in gen-des-tables.c.
We really don't want to use variable-length array *variables* anywhere
in this library, because in the place where people are most likely to
want them - the algorithm modules - the stack size is limited.  So we
want to use -Wvla.  But that also complains about variable-length
array *function parameters* in gen-des-tables.c; this would actually
be totally safe even in the context we're worred about (it's just
syntactic sugar for doing 2D array indexing math yourself) but it's
also not hard to get rid of.  Thus.

	* gen-des-tables.c (write_table_u8, write_table_u32): Don't use
	VLA function parameters.
	(main): Adjust pointer types in calls to above functions.
	* m4/zw_simple_warnings.m4: Reactivate -Wvla.