Improve accuracy of global symbol listings in tests.
In three test scripts, we were using `nm ... | grep ' [A-TV-Z] '` to
get lists of all global symbol definitions in the library.  This is
inaccurate; some uppercase letters are not global (N) and some
lowercase letters *are* global (uvw) and this actually caused a
spurious failure with clang.  It's better to use nm -g; nm actually
knows which symbols are and are not global, and won't get out of
date.  We still have to grep out the undefined symbols, alas.

	* test-symbols-compat, test-symbols-renames, test-symbols-static:
	Use nm -g to list only global symbols.