Install the OpenWall crypt(3) manpage as crypt_rn(3).
We are using the name crypt_rn(3) to avoid a collision with the
crypt(3) page that already exists on most systems.  In this commit,
the text of the manpage is verbatim what it is in crypt_blowfish 1.3;
it will be reworked for consistency with the many changes in this
library in future commits.