Support for cross-compilation in the build scripts.
This involves grubbing around in the guts of Autoconf so it may be a
bit fragile, but there doesn't seem to be a better or more official
way to do it.  There are two difficulties.  When $(CC) is a
cross-compiler, we also need to find a native compiler,
$(CC_FOR_BUILD), to compile gen-des-tables with; Automake doesn't
_really_ support this, and a bunch of configure tests have to be
repeated, but it can be made to work.  Also, the test driver needs to
be taught to treat "exec format error" from the compiled tests as a
skip rather than a fail.

This has only been tested for cross-compilation from x86-linux to
powerpc-linux; in particular there may still be bugs when $(BUILD_OBJEXT)
and $(OBJEXT) are not the same, or when either $(BUILD_EXEEXT) or
$(EXEEXT) is not the empty string.  It may also not work with non-GNU
Make.

	* Makefile.am: Use $(CC_FOR_BUILD) to build gen-des-tables.
	(AM_TESTS_ENVIRONMENT): Add $(CC) to set of exported environment
	variables.
	(LOG_COMPILER): Set to m4/skip-if-exec-format-error when
	cross-compiling.
	(EXTRA_DIST): Add m4/skip-if-exec-format-error.
	* configure.ac: Call AX_PROG_CC_FOR_BUILD.
	Add an AM_CONDITIONAL for $cross_compiling.
	* m4/zw_simple_warnings.m4: When cross-compiling, also determine
	the set of warnings flags usable with $CC_FOR_BUILD.
	* m4/ax_prog_cc_for_build.m4, m4/skip-if-exec-format-error:
	New files.
	* LICENSING, TODO: Update.