Add support for NTHASH ($3$) from FreeBSD
FreeBSD implemented support for the NT LAN Manager hash algorithm to provide
easier compatibility with NT accounts.  The NT-Hash algorithm is known to be
weak, as it uses the deprecated md4 hash algorithm without any salting.
FreeBSD used the $3$ prefix for this.  Its use is not recommended, as it is
easily broken.