Add support for SUNMD5 ($md5[,rounds=%u]$salt$) from (Open)Solaris
This algorithm was developed by Alec Muffett for Solaris, as a
replacement for the aging des_crypt.  It was introduced in Solaris
9u2.  While based on the MD5 message digest, it has very little at
all in common with the md5_crypt algorithm.  It supports 32 bit
variable rounds and an 8 character salt.

The gensalt() function for this algorithm is modified to always
return a new setting with a sane, but random, default amount of
rounds for the corresponding crypt() function.  Besides this it
follows the original implementation from (Open)Solaris.

Its use is not recommended, as today it is easily broken.