crypt-pbkdf1-sha1: Add support for PBKDF1 with SHA1 ($sha1$<rounds>$<salt>$)
PBKDF1 with SHA1, as defined in RFC 2898, section 5.1 applies a hash
function, which is SHA-1, to derive keys.  This implementation is the
modified version from NetBSD using HMAC-SHA1 instead of plain SHA1
hashing, like used in PBKDF2.

PBKDF1 with SHA1 is recommended only for compatibility with existing
applications since the keys it produces may not be large enough for
some applications.