alg-hmac-sha1: Add implementation of HMAC with SHA1
In cryptography, a keyed-hash message authentication code (HMAC) is a
specific type of message authentication code (MAC) involving a
cryptographic hash function and a secret cryptographic key.  It may
be used to simultaneously verify both the data integrity and the
authentication of a message, as with any MAC.

The cryptographic strength of the HMAC depends upon the size of the
secret key that is used.  The most common attack against HMACs is
brute force to uncover the secret key.  HMACs are substantially less
affected by collisions than their underlying hashing algorithms alone.
Therefore, HMAC-SHA1 does not suffer from the same weaknesses that
have been found in SHA1.