crypt: bcrypt should use secure $2b$ hashes by default
As noted in the documentation OpenBSD 5.5 introduced
the `$2b$` prefix for behavior that exactly matches
crypt_blowfish's `$2y$`, and current crypt_blowfish
supports it as well.  Unfortunately, the behavior of
`$2a$` on password characters with the 8th bit set has
to be considered system-specific.

When generating new password hashes, the `$2b$` or
`$2y$` prefix should be used.

If such hashes ever need to be migrated to a system
that does not yet support these new prefixes, the
prefix in migrated copies of the already-generated
hashes may be changed to `$2a$`.