Replace crypt-sunmd5.c with BSD-licensed cleanroom reimplementation.
This eliminates all CDDL-licensed code from the library, as requested
in issue #7 because of GPL incompatibility.

The new implementation was based exclusively on the prose description
of the algorithm in the Passlib v1.7.1 documentation, which should be
adequately arms-length for copyright purposes.  (N.B. the lengthy
quotation from _Hamlet_ is used as input to MD5, so it must remain
byte-for-byte identical to achieve interoperability with the original;
also, the play itself is in the public domain.)

The new implementation is also fully deterministic (that is, its
gensalt procedure draws randomness only from its rbytes argument) and
does not call malloc.