Make crypt_gensalt for $sha1 deterministic.
All of the randomness is now drawn from the 'rbytes' argument, rather
than taking some of it from direct calls to get_random_bytes.  This
simplifies the logic and enables us to test the output more
thoroughly.

Relatedly, the amount of randomness crypt_gensalt will supply to $sha1
and $3 when it calls get_random_bytes itself is no longer oversized.