Make XCRYPT_SECURE_MEMSET() syntactically a normal expression.
All of the alternative definitions of XCRYPT_SECURE_MEMSET expanded to
a full statement including the terminating semicolon, and none of its
uses had a terminating semicolon, which looked like a typo.  Correct
this: put semicolons on each use, and remove them from the
definitions.  This is entirely cosmetic, but it's been bugging me.

(There is no risk of unexpected parser behavior after macro expansion:
after this change, XCRYPT_SECURE_MEMSET(...) looks like a function
call expression, and all of its possible expansions are, in fact,
function call expressions.)