Add more tests based on gaps in line coverage.
We were not adequately testing invalid-argument error paths within
crypt_gensalt, and we were also not adequately testing a variety of
corner cases related to non-default round parameters

After this change, I believe that all of the non-covered lines fall
into three categories: genuinely impossible to hit from the public
API (usually these are double-checks for preconditions already
verified at a higher level); checks for memory allocation failure; and
fallback paths within get_random_bytes.  The first two of these, IMHO,
are fine to leave as is.  It would probably be worth writing a test
that exercises the get_random_bytes fallbacks but it'd involve getting
one's hands dirty with ELF.  I may do it eventually if no one else does.

This change makes us pickier about non-default round parameters to $5$
and $6$ hashes; numbers outside the valid range are now rejected, as
are numbers with leading zeroes and an explicit request for the
default number of rounds.  This is in keeping with the observation, in
the Passlib documentation, that allowing more than one valid crypt
output string for any given (rounds, salt, phrase) triple is asking
for trouble.

This partially reverts commit
a7f9df50cecec46bb8176382faa685ce35ca72be, which accidentally caused us
to fail to reject invalid DES setting strings.  Because of this, a few
lines in crypt-des.c cease to be covered; they are in category
1 (double-checks of preconditions).