Add tests for get_random_bytes.
It turns out not to be *that* hard to exercise the fallback logic in
get_random_bytes, thanks to GNU ld's --wrap feature.  There is also
some basic black-box testing of the get_random_bytes interface.

The change to randombytes.c itself ensures 100% predictable behavior
if get_random_bytes should ever be called with buflen zero.