Correct crypt.5 terminology (issue #29)
Solar Designer points out that a phrase used to describe DES-crypt,
“Hash size 64 bits (effectively 56),” is misleading.  Revert to
separate “hash size” and “effective key size” table entries as is
used in the Openwall version of this manpage, but omit the “effective
key size” table entry if it would be the same number as “hash size.”
Correct the `.hash` line for yescrypt as discussed in the review of
pull request #27.

Also swap two of the entries in the list of attacks that salt defeats,
for better prose flow; reference RFC 4648 when talking about ‘the
common “base64” encoding’ that hashes don’t use; and tighten up the
implementation of the `.hash` macro.