Correct failures caught by 'make distcheck'.
yescrypt-platform.c is not otherwise mentioned in the Makefile, so it
needs to be in EXTRA_DIST to ensure it makes it into tarballs.

Adding .deps/*.Plo to DISTCLEANFILES works around an apparent bug in
automake 1.16 where we get duplicate deps files for the object files
that are both part of libcrypt.la and referenced directly, and only
one set is properly cleaned up.